+++ src/compiler/parser/mod.rs
@@ -22,8 +22,10 @@ mod lines_6000_6500;
 pub use lines_6000_6500::*;
 mod lines_6500_7000;
 pub use lines_6500_7000::*;
-mod lines_7000_8000;
-pub use lines_7000_8000::*;
+mod lines_7000_7500;
+pub use lines_7000_7500::*;
+mod lines_7500_8000;
+pub use lines_7500_8000::*;
 mod lines_8000_9000;
 pub use lines_8000_9000::*;
 mod lines_9000_end;
